refs #9 Fixed a bug - deleting a transcript, checking delete audio file when there is none -> failure. Now works.

+++ ca.mcgill.cs.swevo.qualyzer/src/ca/mcgill/cs/swevo/qualyzer/handlers/DeleteTranscriptHandler.java
@@ -101,7 +101,7 @@ public class DeleteTranscriptHandler extends AbstractHandler
 			//TODO delete codes
 		}
 		
-		if(deleteAudio)
+		if(deleteAudio && transcript.getAudioFile() != null)
 		{
 			File file = new File(wProject.getLocation() + transcript.getAudioFile().getRelativePath());
 			file.delete();
